added getstarted, loading, and success components
