+++ src/components/pages/Analytics.js
@@ -3,6 +3,7 @@ import Stats from "../analytics/Stats";
 import PopularWords from "../analytics/PopularWords";
 
 
+
 const Analytics = () => {
   return (
     <Fragment>
